data

+++ src/app/category/test/test.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { CategoryService } from '../../shared/services/category.service';
 import { Router } from '@angular/router';
+import { delay } from 'rxjs/operators';
 
 @Component({
   selector: 'app-test',
@@ -8,12 +9,10 @@ import { Router } from '@angular/router';
   styleUrls: ['./test.component.scss']
 })
 export class TestComponent implements OnInit {
-  myname="isha";
-  onSubmit(value:any){
-    console.log(value);
-  }
+
   categories: any[] = [];
-  isDataAvailable=false;
+  isDataAvailable = false;
+  visible = false;
   constructor(private categoryservice: CategoryService, private router: Router) {
 
   }
@@ -21,15 +20,26 @@ export class TestComponent implements OnInit {
 
 
   ngOnInit() {
-   this.categoryservice.getcategoriesfromjson().subscribe(data=>{
-     this.categories=data;
-     this.isDataAvailable=true;
-   },
-   error=>{
+    this.getResults();
+    this.visible = true;
+    this.categoryservice.getcategoriesfromjson().pipe(delay(5000)).subscribe(data => {
+      this.categories = data;
+      this.isDataAvailable = true;
+      this.visible = false;
+    },
+      error => {
 
-   
-   });
+
+      });
   }
+
+call(){}
+  getResults() {
+    this.categoryservice.sendRequest().subscribe(data => {
+      console.log(data);
+    })
+  }
+
   navigate(id) {
     this.router.navigate(['test/detail', id])
   }
@@ -37,11 +47,17 @@ export class TestComponent implements OnInit {
   navi(id) {
     this.router.navigate(['editCategory', id])
   }
-  deleteRow(id) {
-    
 
-    if (confirm("Are you sure")) {
-      this.categoryservice.deleteCategoryById(id)
+  deleteRow(id) {
+
+    if (id) {
+
+      if (confirm("Are you sure")) {
+        this.categoryservice.deleteCategoryById(id)
+      }
+      else ("data not found")
+
     }
   }
+
 }
